Move User repo to `repository`
